refactor: GroupMainActivity를 Fragment로 변환하여, 바텀 네비게이션 뷰에 넣어줌
