+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
         </activity>
 
         <activity
-          android:name=".Activity.GroupMainActivity"
+          android:name=".ui.screen.group.MemberListActivity"
           android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -41,25 +41,11 @@
         </activity>
 
         <activity
-          android:name=".Activity.MemberListActivity"
-          android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-
-            <meta-data
-              android:name="android.app.lib_name"
-              android:value="" />
-        </activity>
-
-        <activity
-          android:name=".Activity.GroupCreateActivity"
+          android:name=".ui.screen.group.GroupCreateActivity"
           android:exported="true"/>
 
         <activity
-          android:name=".Activity.GroupInfoActivity"
+          android:name=".ui.screen.group.GroupInfoActivity"
           android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
